Change float to double

+++ src/main/java/com/example/tennis/persistence/entity/GameType.java
@@ -26,6 +26,6 @@ public class GameType {
 
     @NotNull
     @Column(name = "PRICE_MULTIPLIER")
-    private float priceMultiplier;
+    private double priceMultiplier;
 
 }
